fix(power): normalize units to Watts and refactor power types to i32
- Fix unit mismatch where Home Assistant provided kW while scheduler expected Watts\n- Refactor all internal power representations (production, consumption, expected load) to signed integers (i32)\n- Update fetch_ha_state to automatically normalize kW to Watts\n- Maintain kW display in logs while using Watts internally\n- Fix various compiler warnings (unused variables, imports, and mutability) in server and tests\n- Update documentation to mandate Watts as the standard internal unit

+++ lmha-core/src/lib.rs
@@ -63,7 +63,7 @@ pub struct Device {
     pub mqtt_topic: String,
     pub name: String,
     pub is_enabled: bool,
-    pub expected_load: f64,
+    pub expected_load: i32,
     pub current_state: DeviceState,
     pub last_heartbeat: Option<DateTime<Utc>>,
 }
@@ -82,7 +82,7 @@ pub struct Telemetry {
     pub timestamp: DateTime<Utc>,
     pub source: TelemetrySource,
     pub device_id: Option<Uuid>,
-    pub value: f64,
+    pub value: i32,
     pub metadata: Option<serde_json::Value>,
 }
